refactor potential-write-proxy.ts so that unwrapping can be done independently of it using dependencies in util.ts
